Added a proof of concept for notifying users of new artist releases
I was unable to find any existing API's to get new releases, and in order to send the notifications as a push notification we would have to set up a server. So, as a proof of concept I created a notification that uses apple's api to get the current top 10 most played albums and sends them as an in-app notification.