user image screen
I made a user image screen, and installed the camera to do it. Then I had to update some other dependencies, and everything broke. Luckily that was on another branch, not the main.

Errors:
expo camera is not installed on main, and the screen cannot be reached in any way.